% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_Classifier.R
\name{buildPredictor}
\alias{buildPredictor}
\title{Build a RFC and test its performances}
\usage{
buildPredictor(data, status = 10, vset = NULL, n = 10000, m = 3)
}
\arguments{
\item{data}{A matrix or data.frame containing subjects as rows and TMS
parameters as columns. A further column specifying subject diagnosis
can be given to validate the classifier.}

\item{status}{Numeric value specifying subject diagnosis (default
status = 10).}

\item{vset}{An optional data.frame of the same format of the input data.
This will be used as external validation set. The validation set can
be generated from input data, using \code{\link[tmsClassifier]{tmsClassify}}.}

\item{n}{Number of bootstrap sampled trees of the RFC (default n = 10000).
Increasing n improves RFC classification performance, but will sensibly
increase computational time.}

\item{m}{Number of random variables for each tree split (default m = 3).
The suggested value is the squared root of the total number of variables
(the default value is set to 3 since 9 TMS parameter are expected by
the default classifier).}
}
\value{
An object of class \code{\link[randomForest]{randomForest}}.
An optional object of class \code{\link[tmsClassifier]{performance}}
is returned, if vset is not NULL.
}
\description{
Build a RFC from TMS regression parameters and test its
classification performances.
}
\references{
Liaw A, Wiener M. Classification and Regression by randomForest (2002). 
R News, 2(3):18-22. https://doi.org/10.1023/A:1010933404324
}
\seealso{
\code{\link[randomForest]{randomForest}}
}
