% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_Classifier.R
\docType{class}
\name{cross.validation-class}
\alias{cross.validation-class}
\title{S4 class to represent k-fold cross-validation results.}
\description{
S4 class to represent k-fold cross-validation results.
}
\section{Slots}{

\describe{
\item{\code{DvsHC}}{Performances for Disease-Healthy classidication}

\item{\code{FTDvsNonFTD}}{Performances for FTD-NonFTD classidication}

\item{\code{ADvsDLB}}{Performances for AD-DLB classidication}
}}

