% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_Classifier.R
\docType{class}
\name{performance-class}
\alias{performance-class}
\title{S4 class to represent classifier performances.}
\description{
S4 class to represent classifier performances.
}
\section{Slots}{

\describe{
\item{\code{ctable}}{Confusion 2x2 table}

\item{\code{performance}}{List of classifier performance indices}
}}

