% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_Classifier.R
\name{performance}
\alias{performance}
\title{Evaluate classifier performances}
\usage{
performance(obs = NULL, pred = NULL, CT = NULL, y = "0,1")
}
\arguments{
\item{obs}{Vector of observed values.}

\item{pred}{Vector of predicted values.}

\item{CT}{An object of class "table". Alternatively to the vectors of
observed and predicted values, the user could provide a 2x2 contingency
table.}

\item{y}{Outcome encoding. Use "0,1" ("1,2") if the binary encoding is
0 (1) for controls and 1 (2) for cases.}
}
\value{
An object of class \code{\link[tmsClassifier]{performance}}.
}
\description{
Compute performance indices for a binary classifier, from
a vector of observed values and a vector of predictions.
}
