% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_Classifier.R
\name{rmout}
\alias{rmout}
\title{Remove outliers using subject-wise Brier scores}
\usage{
rmout(data, status = 10, k = 5, b = 1)
}
\arguments{
\item{data}{A matrix or data.frame containing subjects as rows and TMS
parameters as columns. A further column specifying subject diagnosis
is required to run RFC cross-validation.}

\item{status}{Numeric value specifying subject diagnosis (default
status = 10).}

\item{k}{Number of cross-validation cycles (default k = 5).}

\item{b}{Maximum Brier score beyond which a subject is considered as
an outlier. By default, b = 1. If b is set to NULL, outliers will be
automatically set to Q3 + 1.5*(Q3 - Q1).}
}
\value{
The input data.frame without outliers.
}
\description{
Detect and remove outliers from a dataset using per subject
Brier scores (Brier, 1950).
}
\references{
Brier GW (1950). Verification of forecasts expressed in terms of probability.
Monthly Weather Review, 78(1):1-3.
https://doi.org/10.1175/1520-0493(1950)078<0001:VOFEIT>2.0.CO;2
}
\seealso{
\code{\link[CMA]{evaluation}}
}
