% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_coefficients.R
\docType{data}
\name{tms.coef}
\alias{tms.coef}
\title{Default TMS regression coefficients}
\format{"tms.coef" is a data.frame of 649 subjects (rows) and 9 TMS regression 
coefficients (columns). Subjects are subdivided into 273 Alzheimer's 
Disease patients (AD), 207 Frontotemporal Dementia patients (FTD), 
67 Dementia with Lewy Bodies patients (DLB), and 147 healthy controls (HC).}
\usage{
tms.coef
}
\description{
Data.frame of TMS regression coefficients derived from 
default TMS data (Benussi, 2017). They include regression coefficients 
over time (t) for short-interval intracortical inhibition 
(SICI ~ bs0 + bs*t), intracortical facilitation (ICF ~ bi0 + bi*t), 
short-latency afferent inhibition (SAI ~ b0 + b1*t + b2*t^2), and 
long-interval intracortical inhibition (LICI ~ a0 + a1*t).
}
\references{
Benussi A, Di Lorenzo F, Dell'Era V, Cosseddu M, Alberici A, Caratozzolo S, 
Cotelli MS, Micheli A, Rozzini L, Depari A, Flammini A, Ponzo V, 
Martorana A, Caltagirone C, Padovani A, Koch G, Borroni B. (2017). 
Transcranial magnetic stimulation distinguishes Alzheimer disease from 
frontotemporal dementia. Neurology, 89(7):665-672.
https://doi.org/10.1212/WNL.0000000000004232
}
