% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_RFC1.R
\docType{data}
\name{tms.rfc1}
\alias{tms.rfc1}
\title{D-HC Random Forest Classifier (RFC1)}
\format{"tms.rfc1" is an object of class \code{\link[randomForest]{randomForest}}.}
\usage{
tms.rfc1
}
\description{
This is the first of three RFCs used by tmsClssifier to
predict subjects' disgnosis on the base of TMS data. The RFC1 classify
subjects in either "Diseased" (AD, FTD, or DLB) or "Healthy" (i.e. the
typical TMS signature of heathy controls). If a subject is classified
as "Diseased", the classification process moves to the second classifier
(RFC2). In case of an "Healthy" subject, the classification process
terminates.
}
