% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_RFC2.R
\docType{data}
\name{TMS_RFC2}
\alias{TMS_RFC2}
\title{FTD-NonFTD Random Forest Classifier (RFC2)}
\format{"tms.rfc2" is an object of class \code{\link[randomForest]{randomForest}}.}
\usage{
tms.rfc2
}
\description{
This is the second of three RFCs used by tmsClssifier to
predict subjects' disgnosis on the base of TMS data. For each RFC1
classification that yielded "Diseased", RFC2 decides if the respective
subject is of "FTD" (Frontotemporal Dementia) or "Non-FTD" class.
In case of an FTD subject, the classification process terminates.
}
