% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_RFC3.R
\docType{data}
\name{TMS_RFC3}
\alias{TMS_RFC3}
\title{AD-DLB Random Forest Classifier (RFC3)}
\format{"tms.rfc3" is an object of class \code{\link[randomForest]{randomForest}}.}
\usage{
tms.rfc3
}
\description{
This is the third of three RFCs used by tmsClssifier to
predict subjects' disgnosis on the base of TMS data. For each RFC2
classification that yielded "Non-FTD", RFC3 decides if the respective
subject is of "AD" (Alzheimer's Disease) or "DLB" (Dementia with Lewy
Bodies) class.
}
