% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_Classifier.R
\name{tmsImpute}
\alias{tmsImpute}
\title{Missing data imputation}
\usage{
tmsImpute(tms, sici.icf = 1:7, sai = 8:11, lici = 12:14,
  max.na = 7, max.r = 0.5, max.c = 0.8, k = 10)
}
\arguments{
\item{tms}{A matrix or data.frame containing subjects as rows and TMS
values as columns.}

\item{sici.icf}{Numeric vector determining the position of
temporally-ordered SICI-ICF columns (SICI: short-interval intracortical
inhibition; ICF: intracortical facilitation). By default, they should
be the first 7 measures (sici.icf = 1:7; 4 for SICI and 3 for ICF),
taken at times (interstimulus intervals): 1, 2, 3, 5, 7, 10, 15 ms.
Set sici.icf to NULL to exclude these values from classification.}

\item{sai}{Numeric vector determining the position of
temporally-ordered SAI (short-latency afferent inhibition) columns.
By default, they should be the 4 columns following sici.icf (sai = 8:11),
taken at time steps (interstimulus intervals): -4, 0, 4, 8 ms. Set sai
to NULL to exclude these values from classification.}

\item{lici}{Numeric vector determining the position of temporally-ordered
LICI (long-interval intracortical inhibition) columns. By default, they
should be the 3 columns following sai (lici = 12:14), taken at time
steps (interstimulus intervals): 50, 100, 150 ms. Set lici to NULL to
exclude these values from classification.}

\item{max.na}{Maximum number of missing values per row (default max.na = 7).}

\item{max.r}{Maximum percentage of missing values per row (default max.r = 0.5).}

\item{max.c}{Maximum percentage of missing values per column (default max.c = 0.8).}

\item{k}{Number of nearest neighbours to perform imputation.}
}
\value{
The original data.frame with imputed missing values.
}
\description{
Impute missing Transcranial Magnetic Stimulation (TMS)
values, using the K-nearest neighbours (KNN) algorithm from
\code{\link[impute]{impute.knn}} (Troyanskaya, 2001).
}
\references{
Troyanskaya O, Cantor M, Sherlock G, Brown P, Hastie T, Tibshirani R,
Botstein D, Altman RB (2001). Missing value estimation methods for
DNA microarrays. Bioinformatics, 17(6):520-525.
https://doi.org/10.1093/bioinformatics/17.6.520
}
\seealso{
\code{\link[impute]{impute.knn}} for further details on KNN
imputation.
}
