% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS_Classifier.R
\name{tmsRegression}
\alias{tmsRegression}
\title{Generate regression parameters from TMS data}
\usage{
tmsRegression(tms, sici.icf = 1:7, sai = 8:11, lici = 12:14,
  adjust = NULL)
}
\arguments{
\item{tms}{A data.frame containing subjects as rows and TMS values as
columns. Optionally, the user may add covariate columns (e.g., sex, age,
center) to adjust TMS regression estimates.}

\item{sici.icf}{Numeric vector determining the position of
temporally-ordered SICI-ICF columns (SICI: short-interval intracortical
inhibition; ICF: intracortical facilitation). By default, they should
be the first 7 measures (sici.icf = 1:7; 4 for SICI and 3 for ICF),
taken at times (interstimulus intervals): 1, 2, 3, 5, 7, 10, 15 ms.
Set sici.icf to NULL to exclude these values from classification.}

\item{sai}{Numeric vector determining the position of
temporally-ordered SAI (short-latency afferent inhibition) columns.
By default, they should be the 4 columns following sici.icf (sai = 8:11),
taken at time steps (interstimulus intervals): -4, 0, 4, 8 ms. Set sai
to NULL to exclude these values from classification.}

\item{lici}{Numeric vector determining the position of temporally-ordered
LICI (long-interval intracortical inhibition) columns. By default, they
should be the 3 columns following sai (lici = 12:14), taken at time
steps (interstimulus intervals): 50, 100, 150 ms. Set lici to NULL to
exclude these values from classification.}

\item{adjust}{Numeric vector determining the position of covariates to
adjust for. By default, adjust = NULL (no covariate adjustment is done).}
}
\value{
A data.frame of estimated regression parameters.
}
\description{
Generate regression parameters from TMS data, needed for
subject classification. For each subject, TMS indicators (SICI-ICF, SAI,
LICI) are modeled as a polynomial functions of time, in the form
y ~ poly(t). This function estimates two parameters for SICI
(SICI = bs0 + bs*t), two parameters for ICF (ICF = bi0 + bi*t), three
parameters for SAI (SAI = b0 + b1*t + b2*t^2), and two parameters
for LICI (LICI = a0 + a1*t).
}
